requirements.txt
fastapi
pytest
uvicorn
nats-py 
cloudevents